# Remove unnecessary sprintf() in hot path