it shows up in blackfire profiles

<img width="447" alt="grafik" src="https://github.com/sebastianbergmann/exporter/assets/120441/d63aee68-4ecc-4dc7-a135-9d86b5801566">