@sebastianbergmann 2024-06-17
Thanks, but as of PHP 8.4 this would not be necessary: https://github.com/php/php-src/pull/14546 ;)