@TimWolla 2024-06-18
FWIW: PHP 8.4 compiles `sprintf()` into string interpolation instead of string concatenation. For 3 segments my benchmark says that they are even, but interpolation gets faster with 5 segments.

The equivalent expression to the PHP 8.4 optimization would be:

```
 "[{$this->shortenedRecursiveExport($data[$key], $processed)}]";
```