@staabm 2024-06-18
thanks for the insight.

> For 3 segments my benchmark says that they are even, but interpolation gets faster with 4 segments.

could you give me an idea what a "segment" is in this context`