@TimWolla 2024-06-18
> could you give me an idea what a "segment" is in this context?

With that I mean the amount of substrings that need to be concatenated. In case of concatenation (`.`) this is the number of `.` plus one. For string interpolation this is the number of variables + the number of literals in between those variables.

The best way to check this is by dumping the Opcodes. You can use the following to dump the Opcodes after Opcache optimization:

```
php -d opcache.enable_cli=1 -d opcache.opt_debug_level=0x20000 test.php
```

I have concat.php:

```
<?php

$a = 'foo';

$len = 0;
for ($i = 0; $i < 100_000_000; $i++) {
    $len += strlen("[" . $a . "-");
}
var_dump($len);
```

and

interpolate.php:

```
<?php

$a = 'foo';

$len = 0;
for ($i = 0; $i < 100_000_000; $i++) {
    $len += strlen("[$a-");
}
var_dump($len);
```

When I execute those, I get:

```
$ time php -d opcache.enable_cli=1 -d opcache.opt_debug_level=0x20000 interpolate.php

$_main:
     ; (lines=16, args=0, vars=3, tmps=4)
     ; (after optimizer)
     ; php-src/interpolate.php:1-10
0000 ASSIGN CV0($a) string("foo")
0001 ASSIGN CV1($len) int(0)
0002 ASSIGN CV2($i) int(0)
0003 JMP 0010
0004 T5 = ROPE_INIT 3 string("[")
0005 T5 = ROPE_ADD 1 T5 CV0($a)
0006 T4 = ROPE_END 2 T5 string("-")
0007 T3 = STRLEN T4
0008 ASSIGN_OP (ADD) CV1($len) T3
0009 PRE_INC CV2($i)
0010 T3 = IS_SMALLER CV2($i) int(100000000)
0011 JMPNZ T3 0004
0012 INIT_FCALL 1 96 string("var_dump")
0013 SEND_VAR CV1($len) 1
0014 DO_ICALL
0015 RETURN int(1)
LIVE RANGES:
     5: 0004 - 0006 (rope)
int(500000000)

________________________________________________________
Executed in    2.23 secs    fish           external
   usr time    2.21 secs  958.00 micros    2.21 secs
   sys time    0.02 secs  321.00 micros    0.02 secs

$ time php -d opcache.enable_cli=1 -d opcache.opt_debug_level=0x20000 concat.php

$_main:
     ; (lines=15, args=0, vars=3, tmps=2)
     ; (after optimizer)
     ; php-src/concat.php:1-10
0000 ASSIGN CV0($a) string("foo")
0001 ASSIGN CV1($len) int(0)
0002 ASSIGN CV2($i) int(0)
0003 JMP 0009
0004 T3 = CONCAT string("[") CV0($a)
0005 T4 = FAST_CONCAT T3 string("-")
0006 T3 = STRLEN T4
0007 ASSIGN_OP (ADD) CV1($len) T3
0008 PRE_INC CV2($i)
0009 T3 = IS_SMALLER CV2($i) int(100000000)
0010 JMPNZ T3 0004
0011 INIT_FCALL 1 96 string("var_dump")
0012 SEND_VAR CV1($len) 1
0013 DO_ICALL
0014 RETURN int(1)
int(500000000)

________________________________________________________
Executed in    2.23 secs    fish           external
   usr time    2.22 secs  830.00 micros    2.22 secs
   sys time    0.01 secs  282.00 micros    0.01 secs
```

Note how the interpolate.php has 3 Opcodes with `ROPE_`. These are the 3 segments. The concat.php has 2 concatenations, resulting in 3 different strings being concatenated, with the runtime being identical.

Now when I adjust the test files to:

```
<?php

$a = 'foo';

$len = 0;
for ($i = 0; $i < 100_000_000; $i++) {
    $len += strlen("[" . $a . "-" . $a . "]");
}
var_dump($len);
```

and

```
<?php

$a = 'foo';

$len = 0;
for ($i = 0; $i < 100_000_000; $i++) {
    $len += strlen("[$a-$a]");
}
var_dump($len);
```

respectively, I get the following:

```
$ time php -d opcache.enable_cli=1 -d opcache.opt_debug_level=0x20000 interpolate.php

$_main:
     ; (lines=18, args=0, vars=3, tmps=5)
     ; (after optimizer)
     ; php-src/interpolate.php:1-10
0000 ASSIGN CV0($a) string("foo")
0001 ASSIGN CV1($len) int(0)
0002 ASSIGN CV2($i) int(0)
0003 JMP 0012
0004 T5 = ROPE_INIT 5 string("[")
0005 T5 = ROPE_ADD 1 T5 CV0($a)
0006 T5 = ROPE_ADD 2 T5 string("-")
0007 T5 = ROPE_ADD 3 T5 CV0($a)
0008 T4 = ROPE_END 4 T5 string("]")
0009 T3 = STRLEN T4
0010 ASSIGN_OP (ADD) CV1($len) T3
0011 PRE_INC CV2($i)
0012 T3 = IS_SMALLER CV2($i) int(100000000)
0013 JMPNZ T3 0004
0014 INIT_FCALL 1 96 string("var_dump")
0015 SEND_VAR CV1($len) 1
0016 DO_ICALL
0017 RETURN int(1)
LIVE RANGES:
     5: 0004 - 0008 (rope)
int(900000000)

________________________________________________________
Executed in    3.16 secs    fish           external
   usr time    3.15 secs    0.00 millis    3.15 secs
   sys time    0.00 secs    1.04 millis    0.00 secs

$ time php -d opcache.enable_cli=1 -d opcache.opt_debug_level=0x20000 concat.php

$_main:
     ; (lines=17, args=0, vars=3, tmps=2)
     ; (after optimizer)
     ; php-src/concat.php:1-10
0000 ASSIGN CV0($a) string("foo")
0001 ASSIGN CV1($len) int(0)
0002 ASSIGN CV2($i) int(0)
0003 JMP 0011
0004 T3 = CONCAT string("[") CV0($a)
0005 T4 = FAST_CONCAT T3 string("-")
0006 T3 = CONCAT T4 CV0($a)
0007 T4 = FAST_CONCAT T3 string("]")
0008 T3 = STRLEN T4
0009 ASSIGN_OP (ADD) CV1($len) T3
0010 PRE_INC CV2($i)
0011 T3 = IS_SMALLER CV2($i) int(100000000)
0012 JMPNZ T3 0004
0013 INIT_FCALL 1 96 string("var_dump")
0014 SEND_VAR CV1($len) 1
0015 DO_ICALL
0016 RETURN int(1)
int(900000000)

________________________________________________________
Executed in    3.98 secs    fish           external
   usr time    3.96 secs  659.00 micros    3.96 secs
   sys time    0.01 secs  224.00 micros    0.01 secs
```

5 Opcodes `ROPE_*` and 4 Opcodes `*CONCAT`, with the concat-based version being measurably slower.

--------------------

> did you say that
> 
> ` "[{$this->shortenedRecursiveExport($data[$key], $processed)}]";`
> 
> is faster then
> 
> `'[' . $this->shortenedRecursiveExport($data[$key], $processed) . ']'`
> 
> ?

I didn't test it with that specific code, but in this case I expect both versions to be equal. It would be different if there was another variable to be inserted.

`sprintf()` in PHP 8.4+ with only `%s` and `%d` placeholders will be compiled into the `ROPE_*` version (with `(int)` casts inserted for `%d`).