@TimWolla 2024-06-18
Internally the different between CONCAT and ROPE is that the ROPE will first collect all segment strings and only at the ROPE_END it will sum up the lengths of the segments strings and copy them into the result string. For CONCAT it will generate additional intermediate strings for each concatenation, resulting in redundant copying of the initial segments.